refactor: port upstream changes for merkle
